Extract GridPointClass to its own module
